requirements.txt
accelerate==1.7.0
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp_socks==0.10.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
breadability==0.1.20
cchardet==2.1.7
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cssselect==1.3.0
distro==1.9.0
docopt==0.6.2
elastic-transport==8.17.1
elasticsearch==9.0.2
exceptiongroup==1.3.0
fake-useragent==2.2.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.5.1
geographiclib==2.0
geopy==2.4.1
googletransx==2.4.2
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
lightgbm==4.6.0
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
networkx==3.2.1
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
openai==1.86.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
psutil==7.0.0
pycares==4.9.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-socks==2.7.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.13.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
snscrape==0.7.0.20230622
soupsieve==2.7
sumy==0.11.0
sympy==1.14.0
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.52.4
twint==2.1.20
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
xgboost==2.1.4
yarl==1.20.1
zipp==3.23.0
